fix(inference): keep ChatGPT credential helpers in a dependency-free module
The CLI's lazy import of the shared ChatGPT subscription helpers pulled in
the connections and database graph, which breaks under the CLI test's
secure-keys mock. Split the OAuth config, credential key constants, and the
CES-only credential writer into chatgpt-subscription-credentials.ts; the
DB-backed connection upsert stays in chatgpt-subscription-auth.ts. Update
the secure-keys importer allowlist to match.

Part of JARVIS-1636
